Fix to int

+++ src/itdelatrisu/opsu/audio/MusicController.java
@@ -222,7 +222,7 @@ public class MusicController {
 		double beatLength = lastTimingPoint.getBeatLength() * 100.0;
 		int beatTime = lastTimingPoint.getTime();
 		if (trackPosition < beatTime)
-			trackPosition += (beatLength / 100.0f) * (beatTime / lastTimingPoint.getBeatLength());
+			trackPosition += (int) ((beatLength / 100.0) * (beatTime / lastTimingPoint.getBeatLength()));
 		return (float) ((((trackPosition - beatTime) * 100.0) % beatLength) / beatLength);
 	}
 
@@ -250,7 +250,7 @@ public class MusicController {
 		double measureLength = lastTimingPoint.getBeatLength() * lastTimingPoint.getMeter() * k * 100.0;
 		int beatTime = lastTimingPoint.getTime();
 		if (trackPosition < beatTime)
-			trackPosition += (measureLength / 100.0f) * (beatTime / lastTimingPoint.getBeatLength());
+			trackPosition += (int) ((measureLength / 100.0) * (beatTime / lastTimingPoint.getBeatLength()));
 		return (float) ((((trackPosition - beatTime) * 100.0) % measureLength) / measureLength);
 	}
